+++ Engine/Math/src/Converter.cpp
@@ -3,4 +3,4 @@
 
 
 // Initialize the static scale variable
-float Math::Converter::scale = 1.0f; // Default scale factor
+double Math::Converter::scale = 1.0f; // Default scale factor
